Cache test dependencies
Since libertem/libertem:continuous changes all the time when the code
changes, we establish a base image only with the base requirements,
FIRST add the test requirements to it, and THEN install LiberTEM.
That allows to re-use the layer with the test requirements.

The production image is still tested for scheduler and workers.

+++ tests/template/test_ring_template.py
@@ -101,12 +101,13 @@ def test_ring_tcp_cluster(lt_ctx, random_hdf5_1, scheduler_addr):
     notebook = io.StringIO(notebook.getvalue())
     nb = nbformat.read(notebook, as_version=4)
     ep = ExecutePreprocessor(timeout=600)
+    data_path = None
     try:
         ep.preprocess(nb, {"metadata": {"path": tmp_dir}})
         data_path = os.path.join(tmp_dir, 'ring_result.npy')
         results = np.load(data_path)
     finally:
-        if os.path.exists(data_path):
+        if data_path is not None and os.path.exists(data_path):
             os.remove(data_path)
 
     analysis = lt_ctx.create_ring_analysis(
